+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `MemcacheCacheHandler`, a cache handler that stores the token info in a flask session.
 - Added support for audiobook endpoints: get_audiobook, get_audiobooks, and get_audiobook_chapters.
 - Added integration tests for audiobook endpoints.
 - Removed `python 2.7` from GitHub Actions CI workflow. Python v2.7 reached end of life support and is no longer supported by Ubuntu 20.04.